requirements.txt
numpy>=2.1.0
pandas>=2.2.0
pyarrow>=18.0.0
pyspark>=3.5.0
pyomo>=6.9.0
highspy>=1.8.0
PyYAML>=6.0.0
matplotlib>=3.9.0
pytest>=8.0.0
